Projects name were changed.
