+++ phase04/Project/FullTextSearch.csproj
@@ -5,6 +5,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <RootNamespace>phase02</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
