+++ phase04/Project/FullTextSearchProject.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "phase02", "phase02.csproj", "{F8207E57-29F6-4BEE-AC15-302D8B416A6C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FullTextSearch", "FullTextSearch.csproj", "{F8207E57-29F6-4BEE-AC15-302D8B416A6C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject", "..\TestProject\TestProject.csproj", "{A94790C6-4ACC-46F3-A0D9-D121524FF413}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FullTextSearch.Test", "..\FullTextSearch.Test\FullTextSearch.Test.csproj", "{A94790C6-4ACC-46F3-A0D9-D121524FF413}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
